Refactor classes to make much greater use of Spring IOC.
